QS: Add flippable tiles
This commit adds support for tiles that can be flipped to reveal
more information, as well as a two flippable tiles:
- A Wi-Fi back tile displaying the current IP address in case
you're connected to a Wi-Fi network. Tapping the back of the
tile opens the tethering settings, if supported.
- A Bluetooth back tile displaying the number of paired devices,
and tapping it toggles discoverability.

This commit also moves the loading spinner to the BasicTile, as
it is only supported in basic_tile layout.

Change-Id: Ic08b08a97c3dd999b1915c7bb108c850a2cbcd9b

Conflicts:

	packages/SystemUI/src/com/android/systemui/statusbar/phone/QuickSettings.java

	packages/SystemUI/src/com/android/systemui/statusbar/phone/QuickSettingsBasicTile.java

	packages/SystemUI/src/com/android/systemui/statusbar/phone/QuickSettingsTileView.java

Conflicts:

	packages/SystemUI/src/com/android/systemui/statusbar/phone/QuickSettings.java